Have a panel to dispaly color temp

+++ internal/homeassistant/lights.go
@@ -76,7 +76,7 @@ func (hc *HaClient) ChangeBrightness(entityID string, brightness uint8) error {
 }
 
 // ChangeColorTemp changes the color temperature of a light entity to the specified value in kelvin.
-func (hc *HaClient) ChangeColorTemp(entityID string, colorTemp float64) error {
+func (hc *HaClient) ChangeColorTemp(entityID string, colorTemp int) error {
 	body := map[string]any{
 		"entity_id":  entityID,
 		"color_temp": colorTemp,
